pyproject.toml
[project]
name = "tracuu"
version = "2.0"
requires-python = "==3.14.2"
dependencies = [
    "Django==6.0",
    "django-storages==1.14.6",
    "gunicorn==23.0.0",
    "boto3==1.42.12",
    "pillow==12.0.0",
    "python-dotenv==1.2.1",
    "djangorestframework==3.16.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit==4.5.1",
    "ruff==0.14.10",
    "black==25.12.0",
]

[tool.ruff]
line-length = 89
exclude = [
    ".git",
    ".github",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
]

[tool.ruff.lint]
select = ["ALL"]
allowed-confusables = []
ignore = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN201",
    "ANN202",
    "BLE001",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D401",
    "DJ008",
    "ERA001",
    "G004",
    "PTH",
    "RUF012",
    "TRY400",
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.pylint]
max-args = 10
